Compiler warning fixes.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8038 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ renderer/GLSLProgram.cpp
@@ -313,7 +313,7 @@ GLuint GLSLProgramLoader::CompileShader( GLint shaderType, const char *sourceFil
 		// unfortunately, GLSL compilers don't reference any actual source files in their errors, but only
 		// file index numbers. So we'll display a short legend which index corresponds to which file.
 		ss << "File indexes:\n";
-		for( int i = 0; i < sourceFiles.size(); ++i ) {
+		for( size_t i = 0; i < sourceFiles.size(); ++i ) {
 			ss << "  " << i << " - " << sourceFiles[i] << "\n";
 		}
 		common->Warning( ss.str().c_str() );
